Update Behaviour added, Replaced BIGNT ID- VALUES with INT VALUES

+++ src/at/ac/tuwien/docspars/entity/TimestampedDict.java
@@ -1,7 +1,5 @@
 package at.ac.tuwien.docspars.entity;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.sql.Timestamp;
 
 import at.ac.tuwien.docspars.io.db.SQLStatements;
@@ -13,7 +11,7 @@ public class TimestampedDict extends Dict {
 	private int docFQ;	
 	private TimestampedDict predescessor;
 	
-	public TimestampedDict(long id, String term, Timestamp added, int doc_fq) {
+	public TimestampedDict(int id, String term, Timestamp added, int doc_fq) {
 		super(id, term);
 		this.addedTimestamp=added;
 //		this.removedTimestamp=removed;
